requirements.txt
fastapi
uvicorn
rembg
pillow
numpy
torch
torchvision
torchaudio
onnxruntime
python-multipart
requests
mysql-connector-python